Updating default hostname to "vm".
We had the default hostname set to "local". This had been found
to cause issues with certain front-end development tools. The
following was reported:

"...when using browsersync for our development, we need to proxy
the vm and the .local address is used by the native Bonjour app
on OSX, making browsersync barely usable."

This change resolves the address conflict issue.