Prevent error or vagrant reload.
In some cases vagrant would fail to reload a vm due to issues
with the postfix configuration. However we actually don't appear
to need or use postfix on the vm and instead route php mail
through mhsendmail to mailhog. As a result postfix can be removed
and the failing scripts can be replaced with a fast `exit 0`
command.